requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
paho-mqtt>=1.6.0
pytz>=2023.3
